Eliminate circular dependency on headers
