+++ src/io/iter_sparse_batchloader.h
@@ -16,6 +16,7 @@
 #include "./inst_vector.h"
 #include "./image_iter_common.h"
 #include "./iter_batchloader.h"
+#include "./iter_sparse.h"
 
 namespace mxnet {
 namespace io {
